+++ test/clj_kondo/main_test.clj
@@ -1416,6 +1416,18 @@ foo/foo ;; this does use the private var
    '({:file "<stdin>", :row 1, :col 11, :level :error, :message "Unparsable libspec: [foo oh-no :as]"})
    (lint! "(require '[foo oh-no :as])" {:linters {:syntax {:level :error}}})))
 
+(deftest varname-syntax-test
+  ;; TODO: linter is triggered twice, why?
+  ;; TODO: clj-kondo/ignore doesn't work yet
+  (assert-submaps2
+   [{:file "<stdin>", :row 2, :col 7, :level :error, :message "Symbols starting or ending with dot (.) are reserved by Clojure: .foo"}
+    {:file "<stdin>", :row 3, :col 7, :level :error, :message "Symbols starting or ending with dot (.) are reserved by Clojure: bar."}]
+   (lint! "
+(defn .foo [])
+(defn bar. [])
+(defn #_:clj-kondo/ignore .dude. [])
+#_:clj-kondo/ignore (defn .dude2 [])" {:linters {:syntax {:level :error}}})))
+
 (deftest call-as-use-test
   (is (empty?
        (lint!